Some refactoring and javadoc
